be/driver_hos_logbook/pyproject.toml
[project]
name = "driver_hos_logbook"
version = "0.1.0"
description = ""
authors = [
    {name = "Abdulrahman Al-Tayeb",email = "aaltayeb@lumofy.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-configurations (>=2.5.1,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "drf-polymorphic (>=2.1.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-queryable-properties (>=1.12.0,<2.0.0)",
    "drf-standardized-errors (>=0.15.0,<0.16.0)",
    "django-otp (>=1.6.1,<2.0.0)",
    "boto3 (>=1.40.24,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "argon2 (>=0.1.10,<0.2.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dateutil (>=2.9.0,<3.0.0)",
    "unidecode (>=1.4.0,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "reportlab (>=4.4.6,<5.0.0)",
    "svglib (>=1.6.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "sib-api-v3-sdk (>=7.6.0,<8.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
# Add Python constraint here
